ai4ef_data_app/python_requirements.txt
openpyxl==3.1.2
pandas==2.2.2
dagster==1.7.0
dagster-webserver==1.7.0
SQLAlchemy==2.0.23
python-dotenv==1.0.0